fix: finally something that works!! Corrected ui and tweaked stuff for the intersection shapes and made all of this working (almost) with the boundary with circles. this is still yet-to-be-improved!

+++ core/base/src/main/kotlin/org/elnix/dragonlauncher/base/model/serializables/IntersectionShape.kt
@@ -30,7 +30,7 @@ public data class IntersectionShape(
     val scale: Float = 1f,
 
     @SerialName("angle")
-    val angle: Int = 0,
+    val angle: Float = 0f,
 
     @SerialName("offset")
     @Serializable(with = OffsetSerializer::class)
